enigma: Add support for Lockscreen clock fonts
After r16, a commit to cleanup the KeyguardClockSwtich and unify it's structure was merged.
Also, this added a new resource into the frameworks core, which allows us to overlay the custom clock font.
So, we need to overlays now to customize this. This makes me really happy because it removes the ugly method that we were doing years ago.

Signed-off-by: RakeshBatra <raakesh.batra@rediffmail.com>